+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2019-07-24
+### Added
+- Add `WeekPicker` component.
+### Changed
+- Change usage of `week.number` translation key to receive argument `week`
+instead of `weekNumber`.
+
 ## [0.1.5] - 2019-07-23
 ### Added
 - Add `onWeekSelect` property to `DatePicker` to determine what should happen
@@ -32,6 +39,7 @@ Semantic UI Form.
 ### Added
 - First version of this project.
 
+[0.2.0]: https://github.com/CodeYellowBV/daycy/releases/tag/0.2.0
 [0.1.5]: https://github.com/CodeYellowBV/daycy/releases/tag/0.1.5
 [0.1.4]: https://github.com/CodeYellowBV/daycy/releases/tag/0.1.4
 [0.1.3]: https://github.com/CodeYellowBV/daycy/releases/tag/0.1.3